+++ src/org/amse/bomberman/util/interpretator/NumLexema.java
@@ -7,11 +7,11 @@ package org.amse.bomberman.util.interpretator;
  */
 public class NumLexema extends Lexema {
     // the meaning of number
-    public final Long value;
+    public final Double value;
 
     // constructor
     public NumLexema(String value) {
 	super(Lexema.Type.NUMBER);
-	this.value = new Long(value);
+	this.value = new Double(value);
     }
 }
